Update activity table with deleted version/project: https://bugzilla.redhat.com/show_bug.cgi?id=1229940
