+++ zanata-war/src/main/resources/messages.properties
@@ -148,12 +148,20 @@ jsf.NoActivityMessage=As you begin to use Zanata, what you have done will show u
 jsf.dashboard.activity.title=Activity
 jsf.dashboard.activity.translate.message=You translated <strong>{0} words</strong> in \
 <a href="{1}">{2}</a>, finishing on &#8220;<a href="{3}">{4}</a>&#8221;
+jsf.dashboard.activity.translate.message.versionDeleted=You translated <strong>{0} words</strong> in \
+<a href="{1}">{2}</a>, finishing on &#8220;<strong>{3}</strong>&#8221;
+jsf.dashboard.activity.translate.message.projectDeleted=You translated <strong>{0} words</strong> in \
+<strong>{1}</strong>, finishing on &#8220;<strong>{2}</strong>&#8221;
 jsf.dashboard.activity.review.message=You reviewed <strong>{0} words</strong> in \
-<a href="{1}">{2}</a>, finishing on &#8220;<a href="{3}">{4}</a>"&#8221;
-jsf.dashboard.activity.uploadSource.message=You uploaded source documents of <strong>{0} words</strong> to \
-<a href="{1}">{2}</a>
-jsf.dashboard.activity.uploadTranslation.message=You uploaded translations of <strong>{0} words</strong> to \
-<a href="{1}">{2}</a>
+<a href="{1}">{2}</a>, finishing on &#8220;<a href="{3}">{4}</a>&#8221;
+jsf.dashboard.activity.review.message.versionDeleted=You reviewed <strong>{0} words</strong> in \
+<a href="{1}">{2}</a>, finishing on &#8220;<strong>{3}</strong>&#8221;
+jsf.dashboard.activity.review.message.projectDeleted=You reviewed <strong>{0} words</strong> in \
+<strong>{1}</strong>, finishing on &#8220;<strong>{2}</strong>&#8221;
+jsf.dashboard.activity.uploadSource.message=You uploaded source documents of <strong>{0} words</strong> to <a href="{1}">{2}</a>
+jsf.dashboard.activity.uploadSource.message.projectDeleted=You uploaded source documents of <strong>{0} words</strong> to <strong>{1}</strong>
+jsf.dashboard.activity.uploadTranslation.message=You uploaded translations of <strong>{0} words</strong> to <a href="{1}">{2}</a>
+jsf.dashboard.activity.uploadTranslation.message.projectDeleted=You uploaded translations of <strong>{0} words</strong> to <strong>{1}</strong>
 jsf.dashboard.activity.lastTranslatedBy.message=Last translated by {0}
 jsf.dashboard.projects.title=Projects
 jsf.dashboard.projects.maintainedProjects.label=Maintained Projects
