+++ zanata-war/src/main/webapp/WEB-INF/layout/dashboard/activity.xhtml
@@ -26,65 +26,7 @@
         <ul id="activity-list" class="list--panel loader__container">
           <ui:repeat value="#{activities}" var="activity">
             <li class="l--pad-all-half bg--higher--hover bg--faint reveal">
-              <p class="l--float-right txt--neutral l--push-left-half">
-                <i aria-hidden="true"
-                  class="i--large #{activityAction.getActivityTypeIconClass(activity)}"
-                  title="#{activityAction.getActivityTitle(activity)}"></i>
-                <span class="is-invisible">#{activityAction.getActivityTitle(activity)}</span>
-              </p>
-
-              <p class="l--push-all-0">
-                <h:outputText escape="false"
-                  value="#{activityAction.getActivityMessage(activity)}"/>
-              </p>
-              <ul class="list--horizontal txt--meta">
-                <li><i aria-hidden="true" class="i i--clock txt--neutral"
-                  title="Time"></i>
-                  <span class="is-invisible">#{msgs['jsf.Time']}</span>
-                  #{activityAction.getHowLongAgoDescription(activity)}
-                </li>
-
-                <ui:param name="docName"
-                  value="#{activityAction.getDocumentName(activity)}"/>
-
-                <s:fragment
-                  rendered="#{docName != null and !docName.isEmpty()}">
-                  <li class="reveal__target">
-                    <i aria-hidden="true" class="i i--document txt--neutral"
-                      title="#{msgs['jsf.Document']}"></i>
-                    <span
-                      class="is-invisible">#{msgs['jsf.Document']}</span>
-                    <a href="#{activityAction.getDocumentUrl(activity)}">#{docName}</a>
-                  </li>
-                </s:fragment>
-
-                <ui:param name="languageName"
-                  value="#{activityAction.getLanguageName(activity)}"/>
-
-                <s:fragment
-                  rendered="#{languageName != null and !languageName.isEmpty()}">
-                  <li class="reveal__target">
-                    <i aria-hidden="true" class="i i--language txt--neutral"
-                      title="#{msgs['jsf.Language']}"></i>
-                    <span
-                      class="is-invisible">#{msgs['jsf.Language']}</span>
-                    <a href="#{activityAction.getDocumentListUrl(activity)}">#{languageName}</a>
-                  </li>
-                </s:fragment>
-
-                <ui:param name="versionName"
-                  value="#{activityAction.getVersionName(activity)}"/>
-                <s:fragment
-                  rendered="#{versionName != null and !versionName.isEmpty()}">
-                  <li class="reveal__target">
-                    <i aria-hidden="true" class="i i--version txt--neutral"
-                      title="#{msgs['jsf.Version']}"><span
-                      class="is-invisible">#{msgs['jsf.Version']}</span></i>
-                    <a href="#{activityAction.getVersionUrl(activity)}"><zanata:version-label version="#{activityEntry.getVersion(activity)}"/></a>
-                  </li>
-                </s:fragment>
-
-              </ul>
+              <zanata:activity-entry value="#{activity}" hideInfo="true"/>
             </li>
           </ui:repeat>
           <s:fragment rendered="#{activityAction.hasMoreActivities()}">
