+++ zanata-war/src/main/webapp/resources/zanata/version-label.xhtml
@@ -26,7 +26,7 @@
   <s:fragment rendered="#{cc.attrs.version.status == 'OBSOLETE'}">
     <span class="txt--neutral">#{cc.attrs.version.slug}
       <s:fragment rendered="#{cc.attrs.showIcon}">
-        <i class="i--small i--right i--archive" title="#{msgs['jsf.iteration.tooltip.obsolete']}"></i>
+        <i class="i--small i--right i--trash" title="#{msgs['jsf.iteration.tooltip.obsolete']}"></i>
       </s:fragment>
     </span>
   </s:fragment>
